requirements.txt
Django==2.2.2
djangorestframework==3.9.4
django-cors-headers==3.0.1
mysqlclient==1.4.2.post1